requirements.txt
Django==1.11.5
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
mysqlclient==1.3.12
olefile==0.44
Pillow==4.2.1
django-filter==1.0.4